+++ app/src/main/java/com/fibelatti/photowidget/configure/PhotoWidgetConfigureState.kt
@@ -16,6 +16,8 @@ data class PhotoWidgetConfigureState(
 
     sealed class Message {
 
+        data object SuggestImport : Message()
+
         data object ImportFailed : Message()
 
         data object TooManyPhotos : Message()